requirements.txt
discord.py
safygiphy
